return and renew item functions

+++ src/LibraryProject/ItemReturn.java
@@ -9,14 +9,13 @@ public class ItemReturn {
     String itemName;
     static long noOfOutStandingDays;
     int libraryCardNumber;
-    int isItemAvailable;
     static int itemId;
     static double itemValue;
     static double fineForReturn;
     static LocalDate dueDate;
     static String itemCategory;
 
-
+    static int newLibraryCardNumber;
     static LocalDate returnDateString= LocalDate.now();
 
 
@@ -35,7 +34,6 @@ public class ItemReturn {
 
 
                 while (rs2.next()) {
-                    //isItemAvailable= rs2.getInt(7);
                     itemValue = rs2.getDouble(6);
                     itemCategory= rs2.getString(2);
                     itemId= rs2.getInt(1);
@@ -47,29 +45,18 @@ public class ItemReturn {
 
                     ResultSet rs3 = stmt3.executeQuery();
                     if(rs3.next()){
-
-                            //noOfOutStandingDays=rs3.getInt(1);
                             dueDate= rs3.getDate(1).toLocalDate();
-                            System.out.println("days "+noOfOutStandingDays);
-
-                            System.out.println("itemdid:"+itemId);
-
                         Period diff = Period.between(dueDate, returnDateString);
-
-
-                       //fineCalculation
                         noOfOutStandingDays = diff.getDays();
-                        System.out.println("days "+noOfOutStandingDays);
+                        System.out.println("Number of outstanding days: "+noOfOutStandingDays);
                         if(noOfOutStandingDays>0)
                             fineForReturn = (noOfOutStandingDays * .10);
                         else fineForReturn=0;
                         if(fineForReturn>itemValue){
                             fineForReturn=itemValue;
                         }
-                        // isItemAvailable=1;
-                        System.out.println("fine"+fineForReturn);
+                        System.out.println("Fine to be paid for this item : %.2f" +fineForReturn);
                         String returnDate =returnDateString.toString();
-                        System.out.println("new reutn date"+returnDate);
 
                         PreparedStatement stmt4 = conn.prepareStatement("insert into selibrary.return(libraryCardNumber, itemId,returnDate,fineForReturn)" + "values(?,?,?,?);");
                         stmt4.setInt(1, libraryCardNumber);
@@ -84,38 +71,38 @@ public class ItemReturn {
                         PreparedStatement stmt6= conn.prepareStatement("delete from issuedItem where itemId=?;");
                         stmt6.setInt(1, itemId);
                         stmt6.executeUpdate();
-                        System.out.println("The"+itemCategory+itemName+" is returned successfully");
+                        System.out.println("The "+itemCategory+ " "+itemName+" is returned successfully");
                         //if this item was requested, delete from request and move to issuedItem
                         try{
-                            PreparedStatement stmt7= conn.prepareStatement("select * from request where itemId=? limit 1;");
+                            PreparedStatement stmt7= conn.prepareStatement("select * from request where itemId=? order by requestTimeStamp limit 1;");
                             stmt7.setInt(1, itemId);
                             ResultSet rs7 = stmt7.executeQuery();
-                            int newLibraryCardNumber=0;
-                            while(rs7.next()){
-                                newLibraryCardNumber= rs7.getInt(1);
-                            }
-                            System.out.println("This item has an outstanding request from user:"+newLibraryCardNumber+ "issuing to them now");
-                            PreparedStatement stmt8= conn.prepareStatement("delete from request where itemId=? and libraryCardNumber=?;");
-                            stmt8.setInt(1, itemId);
-                            stmt8.setInt(2,newLibraryCardNumber);
-                            stmt8.executeUpdate();
 
-                /*make new object of issuedItem and pass newLibraryCardNumber and itemId to move it.
-                IssuedItem issuedItem=new IssuedItem(itemName,newLibraryCardNumber);
-                issuedItem.issueItem();*/
+                            if(rs7.next()){
+                                newLibraryCardNumber= rs7.getInt(1);
+                                System.out.println("This item has an outstanding request from user: " + newLibraryCardNumber + " issuing to them now");
+                                PreparedStatement stmt8= conn.prepareStatement("delete from request where itemId=? and libraryCardNumber=?;");
+                                stmt8.setInt(1, itemId);
+                                stmt8.setInt(2,newLibraryCardNumber);
+                                stmt8.executeUpdate();
+
+                                //make new object of issuedItem and pass newLibraryCardNumber and itemId to issue it.
+                                IssuedItem issuedItem=new IssuedItem(itemName,newLibraryCardNumber);
+                                issuedItem.issueItem();
+                            }
+                            else{
+                                System.out.println("No outstanding request for this item");
+                            }
+
                         }
                         catch (SQLException e){
-                            System.out.println("No outstanding request for this item");
+                            System.out.println(e);
                         }
 
                     }else{
                         System.out.println("This item is not issued");
                         return;
                     }
-
-
-
-
 
                 //try end here
             }
